pyproject.toml
[project]
name = "diet-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.61.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlachemy",
    "uvicorn>=0.40.0",
]
